Added macro for implementing conversion traits on structs that wrap three-rs types.
Really wanted to replace macros in `src/scene.rs` with some macros that
could be used outside of the crate, but unfortunately that didn't workout
because while internal three objects have access to the `node` field,
structs external to the crate do not. This causes problems in `as_ref` and
if that one was going to deviate, we might as well make a single macro for
the special purpose of classes that wrap three-rs objects.